[Java] Added a filter that splits a serialized message into multiple chunks
git-svn-id: https://apps.man.poznan.pl/svn/muscle/muscle/trunk@734 b725f6ce-2418-43e8-b868-1e0ad3d2b257

+++ src/java/muscle/core/conduit/filter/DropFilter.java
@@ -34,9 +34,9 @@ public class DropFilter<E extends Serializable> extends AbstractFilter<E,E> {
 	private int counter;
 	
 	/** @param newInDtSec in seconds */
-	public DropFilter(int newInDtSec) {
+	public DropFilter(double newInDtSec) {
 		super();
-		outRate = newInDtSec;
+		outRate = (int)newInDtSec;
 	}
 
 	protected void apply(Observation<E> subject) {
